requirements.txt
Django

djangorestframework

psycopg2-binary

gunicorn

dj-database-url

django-environ

djangorestframework-simplejwt